Fix compile: type mismatches in VPN code

+++ android/app/src/main/java/com/cdnhunter/app/vpn/XrayBridge.kt
@@ -28,7 +28,7 @@ object XrayBridge {
 
     fun start(configContent: String, tunFd: Int = 0) {
         val ctrl = controller ?: throw IllegalStateException("Call init() first")
-        ctrl.startLoop(configContent, tunFd.toLong())
+        ctrl.startLoop(configContent, tunFd)
         running = true
     }
